--- src/trade/SingleTradeContainer.js
@@ -1,22 +1,0 @@
-import React, { Component } from 'react';
-import shouldPureComponentUpdate from 'react-pure-render/function';
-import { connect } from 'react-redux';
-import immutableChildrenToJS from 'binary-utils/lib/immutableChildrenToJS';
-
-import TradeCard from './TradeCard';
-import singleTradeSelectors from '../trades/singleTradeSelectors';
-
-@connect(singleTradeSelectors)
-export default class SingleTradeContainer extends Component {
-
-	shouldComponentUpdate = shouldPureComponentUpdate;
-
-	render() {
-		return (
-			<TradeCard
-				index={0}
-                {...immutableChildrenToJS(this.props)}
-			/>
-		);
-	}
-}
