+++ src/trade/TradeParamsContainer.js
@@ -1,6 +1,6 @@
 import React, { Component, PropTypes } from 'react';
 import shouldPureComponentUpdate from 'react-pure-render/function';
-import TradeParams from '../../trade-params/TradeParams';
+import TradeParams from '../trade-params/TradeParams';
 import immutableChildrenToJS from 'binary-utils/lib/immutableChildrenToJS';
 
 export default class TradeParamsContainer extends Component {
