+++ src/trade/TradeViewChartContainer.js
