+++ src/trade/TradeXCard.js
