+++ src/trade/TradeXCardContainer.js
