+++ src/trade/browser/LayoutSelector.js
