+++ src/trade/browser/TradesPage.js
@@ -1,0 +1,9 @@
+import React from 'react';
+import WebPage from '../../containers/WebPage';
+import TradesLayout from './TradesLayout';
+
+export default props => (
+	<WebPage>
+		<TradesLayout {...props} />
+	</WebPage>
+);
