--- src/trades/__tests__/TradesContainer-test.js
@@ -1,9 +1,0 @@
-import React from 'react';
-import { expect } from 'chai';
-// import TradesContainer from '../TradesContainer';
-
-describe('TradesContainer', () => {
-    it.skip('should render with no properties', () => {
-        // expect(<TradesContainer />).to.not.throw;
-    });
-});
