--- src/trades/tradeListSelectors.js
@@ -1,38 +1,0 @@
-import { createStructuredSelector } from 'reselect';
-import {
-    currencySelector,
-    ticksSelector,
-    tradesCountSelector,
-    layoutNSelector,
-    themeSelector,
-    feedLicensesSelector,
-    ohlcSelector,
-} from '../_store/directSelectors';
-import {
-    availableContractsSelector,
-    assetsIsOpenSelector,
-    tradesParamsSelector,
-    tradesPipSizeSelector,
-    tradesTradingTimesSelector,
-    tradesPurchaseInfoSelector,
-} from '../trade/TradeSelectors';
-
-export default createStructuredSelector({
-    assetsIsOpen: assetsIsOpenSelector,
-    contracts: availableContractsSelector,
-    currency: currencySelector,
-    contractChartData: state => state.chartData,
-    layoutN: layoutNSelector,
-    theme: themeSelector,
-    licensesForAllSymbol: feedLicensesSelector,
-    paramsList: tradesParamsSelector,
-    pipSizeList: tradesPipSizeSelector,
-    proposalInfoList: state => state.tradesProposalInfo,
-    purchaseInfoList: tradesPurchaseInfoSelector,
-    tradesCount: tradesCountSelector,
-    ticksForAllSymbols: ticksSelector, // not really!!!
-    ohlcForAllSymbols: ohlcSelector,
-    tradingTimeList: tradesTradingTimesSelector,
-    uiStateList: state => state.tradesUIStates,
-    tradeErrorList: state => state.tradesError,
-});
